add support for multiple user interfaces
