+++ setup.py
@@ -3,7 +3,7 @@
 from setuptools import setup, find_packages
 setup(
     name = "pysk",
-    version = "0.0.2",
+    version = "0.1.0",
     packages = find_packages(),
     install_requires = [
         "signalk_client>=0.2,<0.3",
